+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new alias --drbd-diskless to command "r td" to mimic the option from "r c"
+
 ## [1.27.0] - 2025-11-11
 
 ### Added