+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.27.1] - 2025-12-11
+
 ### Added
 
 - Added new alias --drbd-diskless to command "r td" to mimic the option from "r c"